requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic==2.9.2
python-jose[cryptography]>=3.3.0
PyJWT>=2.9.0
python-multipart>=0.0.9
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4
requests>=2.32.3
